Enable unlimit scanning and notify function
Enabling unlimitted scanning by setting TimeLimit to a negative value
and setting Notify to a function, consuming a Discovered struct, will
let the peerdiscovery library run forever. Each discovery, regardless of
wether known before or not, will be passed to the Notify function. This
implementation is oriented by the observer pattern.