Reintroduce Loopback Device Support
The `filterInterfaces` function reduces the useable network interfaces
for peer discovery. Due to two changes introduced in v1.6.4[^0] and
v1.6.5[^1], a self-discovery over a loopback device was no longer
possible.

For reasons, a device either holds the loopback flag or the multicast
flag, not both together. This commit reflects this.

Furthermore, some refactoring of the `filterInterfaces` function was
performed to ease its understandability.

[^0]: 3ea067dcb02884cd07a07afb3e2719bfb6d1a8f1
[^1]: 78d5e82683c4f79198de5d39b610bf7e25b45a0a